Move OfflineStore into Provider
