+++ sdk/python/feast/infra/provider.py
@@ -2,11 +2,26 @@ import abc
 from datetime import datetime
 from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Union
 
+import pandas
+import pyarrow
+
+from feast.data_source import DataSource
 from feast.feature_table import FeatureTable
 from feast.feature_view import FeatureView
 from feast.protos.feast.types.EntityKey_pb2 import EntityKey as EntityKeyProto
 from feast.protos.feast.types.Value_pb2 import Value as ValueProto
 from feast.repo_config import RepoConfig
+
+ENTITY_DF_EVENT_TIMESTAMP_COL = "event_timestamp"
+
+
+class RetrievalJob(abc.ABC):
+    """RetrievalJob is used to manage the execution of a historical feature retrieval"""
+
+    @abc.abstractmethod
+    def to_df(self):
+        """Return dataset as Pandas DataFrame synchronously"""
+        pass
 
 
 class Provider(abc.ABC):
@@ -22,6 +37,7 @@ class Provider(abc.ABC):
         Reconcile cloud resources with the objects declared in the feature repo.
 
         Args:
+            project: Project to which tables belong
             tables_to_delete: Tables that were deleted from the feature repo, so provider needs to
                 clean up the corresponding cloud resources.
             tables_to_keep: Tables that are still in the feature repo. Depending on implementation,
@@ -39,6 +55,7 @@ class Provider(abc.ABC):
         Tear down all cloud resources for a repo.
 
         Args:
+            project: Feast project to which tables belong
             tables: Tables that are declared in the feature repo.
         """
         ...
@@ -69,6 +86,34 @@ class Provider(abc.ABC):
                 the online store. Can be used to display progress.
         """
         ...
+
+    @staticmethod
+    @abc.abstractmethod
+    def pull_latest_from_table_or_query(
+        data_source: DataSource,
+        entity_names: List[str],
+        feature_names: List[str],
+        event_timestamp_column: str,
+        created_timestamp_column: Optional[str],
+        start_date: datetime,
+        end_date: datetime,
+    ) -> pyarrow.Table:
+        """
+        Note that entity_names, feature_names, event_timestamp_column, and created_timestamp_column
+        have all already been mapped back to column names of the source table
+        and those column names are the values passed into this function.
+        """
+        pass
+
+    @staticmethod
+    @abc.abstractmethod
+    def get_historical_features(
+        config: RepoConfig,
+        feature_views: List[FeatureView],
+        feature_refs: List[str],
+        entity_df: Union[pandas.DataFrame, str],
+    ) -> RetrievalJob:
+        pass
 
     @abc.abstractmethod
     def online_read(
@@ -102,3 +147,31 @@ def get_provider(config: RepoConfig) -> Provider:
         return LocalSqlite(config.online_store.local)
     else:
         raise ValueError(config)
+
+
+def _get_requested_feature_views_to_features_dict(
+    feature_refs: List[str], feature_views: List[FeatureView]
+) -> Dict[FeatureView, List[str]]:
+    """Create a dict of FeatureView -> List[Feature] for all requested features"""
+
+    feature_views_to_feature_map = {}  # type: Dict[FeatureView, List[str]]
+    for ref in feature_refs:
+        ref_parts = ref.split(":")
+        feature_view_from_ref = ref_parts[0]
+        feature_from_ref = ref_parts[1]
+        found = False
+        for feature_view_from_registry in feature_views:
+            if feature_view_from_registry.name == feature_view_from_ref:
+                found = True
+                if feature_view_from_registry in feature_views_to_feature_map:
+                    feature_views_to_feature_map[feature_view_from_registry].append(
+                        feature_from_ref
+                    )
+                else:
+                    feature_views_to_feature_map[feature_view_from_registry] = [
+                        feature_from_ref
+                    ]
+
+        if not found:
+            raise ValueError(f"Could not find feature view from reference {ref}")
+    return feature_views_to_feature_map
